Finalização serviço de login

+++ CodeHubGarage.Domain/JwtSettings.cs
@@ -7,6 +7,7 @@ namespace CodeHubGarage.Domain
         public string Key { get; set; }
         public string Issuer { get; set; }
         public string Audience { get; set; }
-        public double DurationInMinutes { get; set; }
+        public int DurationInMinutes { get; set; }
     }
+
 }
